# Fix a bunch of HTTP measurement and handling issues