This should fix https://github.com/loadimpact/k6/issues/1041, fix https://github.com/loadimpact/k6/issues/1044, fix https://github.com/loadimpact/k6/issues/925, and also fix a potential minor data race in the `httpext.Tracer`! Hopefully, without introducing new bugs or performance regressions...

While investigating this, I realized that we should up the priority of fixing the HTTP digest authentication: https://github.com/loadimpact/k6/issues/800