@na-- 2019-06-10
Benchmarks comparing this branch and master with the same test script from https://github.com/loadimpact/k6/issues/794#issuecomment-500331683:

#### This PR:
```
    data_received..............: 14 GB  232 MB/s
    data_sent..................: 26 MB  428 kB/s
    http_req_blocked...........: avg=250.72µs min=1.6µs   med=2.8µs    max=782.86ms p(90)=4.08µs   p(95)=5.74µs  
    http_req_connecting........: avg=26.95µs  min=0s      med=0s       max=37.43ms  p(90)=0s       p(95)=0s      
    http_req_duration..........: avg=39.39ms  min=84.72µs med=19.08ms  max=506.28ms p(90)=104.04ms p(95)=137.57ms
    http_req_receiving.........: avg=703.28µs min=12.12µs med=45.1µs   max=366.25ms p(90)=320.65µs p(95)=1.35ms  
    http_req_sending...........: avg=38.14µs  min=4.89µs  med=8.94µs   max=39.07ms  p(90)=20.32µs  p(95)=83.25µs 
    http_req_tls_handshaking...: avg=0s       min=0s      med=0s       max=0s       p(90)=0s       p(95)=0s      
    http_req_waiting...........: avg=38.65ms  min=42.43µs med=18.36ms  max=481.31ms p(90)=102.76ms p(95)=135.94ms
    http_reqs..................: 260000 4247.107964/s
    iteration_duration.........: avg=305.41ms min=16.31ms med=290.03ms max=1.16s    p(90)=473.71ms p(95)=531.21ms
    iterations.................: 10000  163.350306/s
    vus........................: 50     min=50 max=50
    vus_max....................: 50     min=50 max=50

57.29user 14.28system 1:03.54elapsed 112%CPU (0avgtext+0avgdata 865536maxresident)k
0inputs+0outputs (0major+222838minor)pagefaults 0swaps
/usr/bin/time k6 run test_case.js  57.30s user 14.29s system 112% cpu 1:03.55 total
```

`master`:
```
    data_received..............: 14 GB  220 MB/s
    data_sent..................: 26 MB  406 kB/s
    http_req_blocked...........: avg=337.5µs  min=1.61µs  med=2.81µs   max=717.91ms p(90)=4.1µs    p(95)=5.85µs  
    http_req_connecting........: avg=34.59µs  min=0s      med=0s       max=83.22ms  p(90)=0s       p(95)=0s      
    http_req_duration..........: avg=41.08ms  min=54.52µs med=18.91ms  max=597.72ms p(90)=109.88ms p(95)=147.07ms
    http_req_receiving.........: avg=41.94µs  min=5.2µs   med=14.38µs  max=42.37ms  p(90)=56.36µs  p(95)=94.92µs 
    http_req_sending...........: avg=41.91µs  min=4.98µs  med=8.96µs   max=51.18ms  p(90)=20.83µs  p(95)=83.65µs 
    http_req_tls_handshaking...: avg=0s       min=0s      med=0s       max=0s       p(90)=0s       p(95)=0s      
    http_req_waiting...........: avg=41ms     min=36.45µs med=18.82ms  max=597.69ms p(90)=109.79ms p(95)=146.97ms
    http_reqs..................: 260000 4026.740937/s
    iteration_duration.........: avg=322.15ms min=13.26ms med=305.36ms max=1.04s    p(90)=508.75ms p(95)=579.46ms
    iterations.................: 10000  154.874651/s
    vus........................: 50     min=50 max=50
    vus_max....................: 50     min=50 max=50

58.94user 14.59system 1:07.19elapsed 109%CPU (0avgtext+0avgdata 844524maxresident)k
0inputs+0outputs (0major+217122minor)pagefaults 0swaps
/usr/bin/time k6 run test_case.js  58.95s user 14.60s system 109% cpu 1:07.20 total
```

The slightly better results for the code of this PR are probably a fluke, since I'm not doing anything that could speed the process up. Quite the opposite, these fixes add 2 mutex operations for each HTTP request... But it probably means that there aren't any huge performance regressions with the new code.