@na-- 2019-06-11
>  Can you add the benchmark as well, as a benchmark :) possibly hitting the local httpbin .

It's going to be a bit tricky to do, but I'll see if I can't work it in. The Go benchmarking logic should hopefully somewhat compensate for the lack of the multiple VUs and no k6 scheduling, though I'm not sure about the buffer pools. It think that benchmark will probably return more reliable results, but they won't be as realistic as the ones from the full end-to-end test.