@na-- 2019-06-13
Not ideal, but... https://twitter.com/CodeWisdom/status/1136250570455605249
```
benchmark                                      old ns/op     new ns/op     delta
BenchmarkHandlingOfResponseBodies/text-4       88628185      90169724      +1.74%
BenchmarkHandlingOfResponseBodies/binary-4     88821807      90402150      +1.78%
BenchmarkHandlingOfResponseBodies/none-4       87605321      88749469      +1.31%

benchmark                                      old allocs     new allocs     delta
BenchmarkHandlingOfResponseBodies/text-4       255475         255597         +0.05%
BenchmarkHandlingOfResponseBodies/binary-4     256182         256332         +0.06%
BenchmarkHandlingOfResponseBodies/none-4       255493         255517         +0.01%

benchmark                                      old bytes     new bytes     delta
BenchmarkHandlingOfResponseBodies/text-4       21949431      21784999      -0.75%
BenchmarkHandlingOfResponseBodies/binary-4     20212931      21772523      +7.72%
BenchmarkHandlingOfResponseBodies/none-4       20217637      20123372      -0.47%
```